requirements.txt
fastapi==0.62.0
pydantic==1.8.1
pandas==1.2.3
pyjwt==1.7.1
